Split game actions into separate files
