+++ Common/Inputs/GameActions/BookInventory.cs
@@ -1,0 +1,22 @@
+﻿namespace untitledplantgame.Common.Inputs.GameActions;
+
+public class BookInventory
+{
+	public const string Left = "book_left";
+	public const string Right = "book_right";
+	public const string Up = "book_up";
+	public const string Down = "book_down";
+
+	public const string North = "book_north";
+	public const string South = "book_south";
+	public const string East = "book_east";
+	public const string West = "book_west";
+
+	public const string TriggerLeft = "book_trigger_left";
+	public const string TriggerRight = "book_trigger_right";
+
+	public const string Confirm = South;
+	public const string CloseBook = North;
+	public const string Back = East;
+	public const string Special = West;
+}
