+++ Common/Inputs/GameActions/Chest.cs
@@ -1,0 +1,21 @@
+﻿namespace untitledplantgame.Common.Inputs.GameActions;
+
+public abstract class Chest
+{
+	public const string Left = "chest_left";
+	public const string Right = "chest_right";
+	public const string Up = "chest_up";
+	public const string Down = "chest_down";
+
+	public const string North = "chest_north";
+	public const string South = "chest_south";
+	public const string East = "chest_east";
+	public const string West = "chest_west";
+
+	public const string TriggerLeft = "chest_trigger_left";
+	public const string TriggerRight = "chest_trigger_right";
+
+	public const string Confirm = South;
+	public const string CloseChest = East;
+	public const string PickUp = West;
+}
