+++ Common/Inputs/GameActions/Dialogue..cs
@@ -1,0 +1,19 @@
+﻿namespace untitledplantgame.Common.Inputs.GameActions;
+
+public abstract class Dialogue
+{
+	public const string Left = "dialogue_left";
+	public const string Right = "dialogue_right";
+	public const string Up = "dialogue_up";
+	public const string Down = "dialogue_down";
+
+	public const string North = "dialogue_north";
+	public const string South = "dialogue_south";
+	public const string East = "dialogue_east";
+	public const string West = "dialogue_west";
+
+	public const string TriggerLeft = "dialogue_trigger_left";
+	public const string TriggerRight = "dialogue_trigger_right";
+
+	public const string Confirm = South;
+}
