+++ Common/Inputs/GameActions/Shop.cs
@@ -1,0 +1,17 @@
+﻿namespace untitledplantgame.Common.Inputs.GameActions;
+
+public class Shop
+{
+	public const string Left = "shop_left";
+	public const string Right = "shop_right";
+	public const string Up = "shop_up";
+	public const string Down = "shop_down";
+	public const string North = "shop_north";
+	public const string South = "shop_south";
+	public const string East = "shop_east";
+	public const string West = "shop_west";
+	public const string TriggerLeft = "shop_trigger_left";
+	public const string TriggerRight = "shop_trigger_right";
+	public const string Confirm = South;
+	public const string CloseShop = East;
+}
